factor out api into a directory
